refactor: directory정리 및 string 분리
