+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.DoSoptTemplate"
         tools:targetApi="31">
         <activity
-            android:name=".login.LoginActivity"
+            android:name=".presentation.login.LoginActivity"
             android:exported="true"
             android:windowSoftInputMode="adjustResize">
             <intent-filter>
@@ -26,10 +26,10 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".SignUpActivity"
+            android:name=".presentation.signup.SignUpActivity"
             android:exported="false" />
         <activity
-            android:name=".home.HomeActivity"
+            android:name=".presentation.home.HomeActivity"
             android:exported="false" />
     </application>
 </manifest>
