--- app/src/main/java/org/sopt/dosopttemplate/HumanAdapter.kt
@@ -1,64 +1,0 @@
-package org.sopt.dosopttemplate
-
-import android.content.Context
-import android.view.LayoutInflater
-import android.view.ViewGroup
-import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView.ViewHolder
-import org.sopt.dosopttemplate.databinding.ItemBirthdayFriendBinding
-import org.sopt.dosopttemplate.databinding.ItemFriendBinding
-import org.sopt.dosopttemplate.databinding.ItemMyBinding
-import org.sopt.dosopttemplate.model.HumanModel
-import org.sopt.dosopttemplate.util.ItemDiffCallback
-
-class HumanAdapter(context: Context) : ListAdapter<HumanModel, ViewHolder>(
-    HomeDiffCallback
-) {
-    private val inflater by lazy { LayoutInflater.from(context) }
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int) =
-        when (viewType) {
-            R.layout.item_my -> {
-                val binding = ItemMyBinding.inflate(inflater, parent, false)
-                MyViewHolder(binding)
-            }
-
-            R.layout.item_friend -> {
-                val binding = ItemFriendBinding.inflate(inflater, parent, false)
-                FriendViewHolder(binding)
-            }
-
-            else -> {
-                val binding = ItemBirthdayFriendBinding.inflate(inflater, parent, false)
-                FriendBirthdayViewHolder(binding)
-            }
-
-        }
-
-    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        val item = currentList[position]//humanList[position]
-
-        when (holder) {
-            is MyViewHolder -> holder.onBind(item as HumanModel.MyModel)
-            is FriendViewHolder -> holder.onBind(item as HumanModel.FriendModel)
-            is FriendBirthdayViewHolder -> holder.onBind(item as HumanModel.FriendBirthdayModel)
-        }
-    }
-
-    override fun getItemCount() = currentList.size
-
-    override fun getItemViewType(position: Int) =
-        when (currentList[position]) {
-            is HumanModel.MyModel -> R.layout.item_my
-            is HumanModel.FriendModel -> R.layout.item_friend
-            is HumanModel.FriendBirthdayModel -> R.layout.item_birthday_friend
-        }
-
-    companion object {
-        private val HomeDiffCallback =
-            ItemDiffCallback<HumanModel>(
-                onItemsTheSame = { old, new -> old.userId == new.userId },
-                onContentsTheSame = { old, new -> old == new }
-            )
-    }
-}
