+++ app/src/main/java/org/sopt/dosopttemplate/presentation/home/user/UserViewModel.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate
+package org.sopt.dosopttemplate.presentation.home.user
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
