+++ app/src/main/java/org/sopt/dosopttemplate/presentation/login/LoginActivity.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate.login
+package org.sopt.dosopttemplate.presentation.login
 
 import android.app.Activity
 import android.content.Intent
@@ -6,15 +6,15 @@ import android.os.Bundle
 import androidx.activity.result.ActivityResult
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.contract.ActivityResultContracts
-import org.sopt.dosopttemplate.ServicePool.authService
-import org.sopt.dosopttemplate.SignUpActivity
+import org.sopt.dosopttemplate.R
 import org.sopt.dosopttemplate.base.BaseActivity
 import org.sopt.dosopttemplate.databinding.ActivityLoginBinding
-import org.sopt.dosopttemplate.home.HomeActivity
+import org.sopt.dosopttemplate.model.User
 import org.sopt.dosopttemplate.model.requestModel.RequestLoginDto
 import org.sopt.dosopttemplate.model.responseModel.ResponseLoginDto
-import org.sopt.dosopttemplate.model.User
-import org.sopt.dosopttemplate.util.getParcelable
+import org.sopt.dosopttemplate.presentation.home.HomeActivity
+import org.sopt.dosopttemplate.presentation.signup.SignUpActivity
+import org.sopt.dosopttemplate.server.ServicePool.authService
 import org.sopt.dosopttemplate.utilprivate.makeToast
 import retrofit2.Call
 import retrofit2.Response
@@ -49,6 +49,42 @@ class LoginActivity : BaseActivity<ActivityLoginBinding>({ ActivityLoginBinding.
         binding.etvLoginPw.setText(pw)
     }
 
+    private fun loginBtn() {
+        binding.btnLoginNaviLogIn.setOnClickListener {
+            val id = binding.etvLoginId.text.toString()
+            val password = binding.etvLoginPw.text.toString()
+
+            authService.login(RequestLoginDto(id, password))
+                .enqueue(object : retrofit2.Callback<ResponseLoginDto> {
+                    override fun onResponse(
+                        call: Call<ResponseLoginDto>,
+                        response: Response<ResponseLoginDto>,
+                    ) {
+                        when (response.isSuccessful) {
+                            true -> {
+                                val data: ResponseLoginDto = response.body()!!
+                                val userId = data.id
+                                val username = data.username
+                                val nickname = data.nickname
+                                val user =
+                                    User(id = userId, username = username, nickname = nickname)
+
+                                moveHomeActivity(user)
+                            }
+
+                            false -> {
+                                makeToast(this@LoginActivity, getString(R.string.USER_INFO_ERROR))
+                            }
+                        }
+                    }
+
+                    override fun onFailure(call: Call<ResponseLoginDto>, t: Throwable) {
+                        makeToast(this@LoginActivity, getString(R.string.SERVER_ERROR))
+                    }
+                })
+        }
+    }
+
     private fun moveHomeActivity(user: User) =
         Intent(this, HomeActivity::class.java).apply {
             putExtra("USER", user)
@@ -67,35 +103,4 @@ class LoginActivity : BaseActivity<ActivityLoginBinding>({ ActivityLoginBinding.
                 this
             }
         )
-
-    private fun loginBtn() {
-        binding.btnLoginNaviLogIn.setOnClickListener {
-            val id = binding.etvLoginId.text.toString()
-            val password = binding.etvLoginPw.text.toString()
-
-            authService.login(RequestLoginDto(id, password))
-                .enqueue(object : retrofit2.Callback<ResponseLoginDto> {
-                    override fun onResponse(
-                        call: Call<ResponseLoginDto>,
-                        response: Response<ResponseLoginDto>,
-                    ) {
-                        if (response.isSuccessful) {
-                            val data: ResponseLoginDto = response.body()!!
-                            val userId = data.id
-                            val username = data.username
-                            val nickname = data.nickname
-                            val user = User(id = userId, username = username, nickname = nickname)
-
-                            makeToast(this@LoginActivity, "로그인이 성공하였고 유저의 ID는 $userId 입니둥")
-
-                            moveHomeActivity(user)
-                        }
-                    }
-
-                    override fun onFailure(call: Call<ResponseLoginDto>, t: Throwable) {
-                        makeToast(this@LoginActivity, "서버 에러 발생")
-                    }
-                })
-        }
-    }
 }
