+++ app/src/main/java/org/sopt/dosopttemplate/server/ApiFactory.kt
@@ -1,14 +1,17 @@
-package org.sopt.dosopttemplate
+package org.sopt.dosopttemplate.server
 
-import android.util.Log
+import  android.util.Log
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import kotlinx.serialization.json.Json
 import okhttp3.Interceptor
 import okhttp3.MediaType.Companion.toMediaType
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
-import org.sopt.dosopttemplate.ApiFactory.BASE_URL
-import org.sopt.dosopttemplate.ApiFactory.USER_BASE_URL
+import org.sopt.dosopttemplate.BuildConfig
+import org.sopt.dosopttemplate.server.ApiFactory.BASE_URL
+import org.sopt.dosopttemplate.server.ApiFactory.USER_BASE_URL
+import org.sopt.dosopttemplate.server.service.AuthService
+import org.sopt.dosopttemplate.server.service.UserService
 import retrofit2.Retrofit
 
 object ApiFactory {
@@ -34,15 +37,8 @@ object ApiFactory {
             .addConverterFactory(Json.asConverterFactory("application/json".toMediaType()))
             .build()
 
-//    val retrofit: Retrofit by lazy{
-//        Retrofit.Builder()
-//            .baseUrl(BASE_URL)
-//            .client(okHttpClient)
-//            .addConverterFactory(Json.asConverterFactory("application/json".toMediaType()))
-//            .build()
-//    }
-
-    inline fun <reified T, B> create(url: B): T = getRetrofit(url.toString()).create<T>(T::class.java)
+    inline fun <reified T, B> create(url: B): T =
+        getRetrofit(url.toString()).create<T>(T::class.java)
 }
 
 object ServicePool {
