+++ app/src/main/res/layout/activity_my_page.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:paddingHorizontal="30dp"
     android:paddingVertical="60dp"
-    tools:context=".login.LoginActivity">
+    tools:context=".presentation.login.LoginActivity">
 
     <ImageView
         android:id="@+id/iv_main_user"
